Added __floordiv__ operator
