+++ src/core/expr/fbinary/bimaker.h
@@ -81,7 +81,7 @@ bimaker_ptr resolve_op(Op, SType, SType);
 // bimaker_ptr resolve_op_minus(SType, SType);
 // bimaker_ptr resolve_op_multiply(SType, SType);
 // bimaker_ptr resolve_op_divide(SType, SType);
-bimaker_ptr resolve_op_intdiv(SType, SType);
+// bimaker_ptr resolve_op_intdiv(SType, SType);
 bimaker_ptr resolve_op_modulo(SType, SType);
 bimaker_ptr resolve_op_power(SType, SType);
 bimaker_ptr resolve_op_and(SType, SType);
