+++ src/core/expr/fbinary/fexpr__truediv__.cc
@@ -25,7 +25,7 @@ namespace dt {
 namespace expr {
 
 
-class FExpr_BinaryDivide : public FExpr_BinaryOp {
+class FExpr__truediv__ : public FExpr_BinaryOp {
   public:
     using FExpr_BinaryOp::FExpr_BinaryOp;
     using FExpr_BinaryOp::lhs_;
@@ -81,7 +81,7 @@ class FExpr_BinaryDivide : public FExpr_BinaryOp {
 
 py::oobj PyFExpr::nb__truediv__(py::robj lhs, py::robj rhs) {
   return PyFExpr::make(
-            new FExpr_BinaryDivide(as_fexpr(lhs), as_fexpr(rhs)));
+            new FExpr__truediv__(as_fexpr(lhs), as_fexpr(rhs)));
 }
